Splitted up socket server and api on two separate servers Changed authorization to role based using JWT token Refactored errors Added feedback service

+++ src/main/java/com/hedbanz/hedbanzAPI/service/Implementation/PasswordResetServiceImpl.java
@@ -2,6 +2,8 @@ package com.hedbanz.hedbanzAPI.service.Implementation;
 
 import com.hedbanz.hedbanzAPI.entity.PasswordResetKeyWord;
 import com.hedbanz.hedbanzAPI.entity.User;
+import com.hedbanz.hedbanzAPI.error.InputError;
+import com.hedbanz.hedbanzAPI.error.NotFoundError;
 import com.hedbanz.hedbanzAPI.error.PasswordResetError;
 import com.hedbanz.hedbanzAPI.exception.ExceptionFactory;
 import com.hedbanz.hedbanzAPI.model.Mail;
@@ -15,13 +17,13 @@ import org.apache.http.util.TextUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.MessageSource;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Random;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -35,33 +37,37 @@ public class PasswordResetServiceImpl implements PasswordResetService{
     private final UserRepository userRepository;
     private final PasswordResetKeyWordRepository keyWordRepository;
     private final EmailService emailService;
+    private final PasswordEncoder passwordEncoder;
 
     private final static String[] LOCALES = {"en", "ru", "uk"};
 
     @Autowired
     public PasswordResetServiceImpl(
-            @Qualifier("MyMessageSource")MessageSource messageSource, UserRepository userRepository,
-            PasswordResetKeyWordRepository keyWordRepository, EmailService emailService) {
+            @Qualifier("MyMessageSource") MessageSource messageSource, UserRepository userRepository,
+            PasswordResetKeyWordRepository keyWordRepository, EmailService emailService, PasswordEncoder passwordEncoder) {
         this.messageSource = messageSource;
         this.userRepository = userRepository;
         this.keyWordRepository = keyWordRepository;
         this.emailService = emailService;
+        this.passwordEncoder = passwordEncoder;
     }
 
     @Transactional
     public void generatePasswordResetKeyWord(PasswordResetData passwordResetData){
         if(TextUtils.isEmpty(passwordResetData.getLogin()))
-            throw ExceptionFactory.create(PasswordResetError.EMPTY_LOGIN);
+            throw ExceptionFactory.create(InputError.EMPTY_LOGIN);
         if(TextUtils.isEmpty(passwordResetData.getLocale()))
-            throw ExceptionFactory.create(PasswordResetError.EMPTY_LOCALE);
+            throw ExceptionFactory.create(InputError.EMPTY_LOCALE);
         if(!isValidLocale(passwordResetData.getLocale()))
-            throw ExceptionFactory.create(PasswordResetError.INCORRECT_LOCALE);
+            throw ExceptionFactory.create(InputError.INCORRECT_LOCALE);
 
         PasswordResetKeyWord oldKeyWord = keyWordRepository.findByUser_login(passwordResetData.getLogin());
         if(oldKeyWord != null)
             keyWordRepository.delete(oldKeyWord);
 
         User user = getUserByLoginOrEmail(passwordResetData.getLogin());
+        if(user == null)
+            throw ExceptionFactory.create(NotFoundError.NO_SUCH_USER);
 
         PasswordResetKeyWord newKeyWord = new PasswordResetKeyWord();
         newKeyWord.setKeyWord(KeyWordGenerator.getGeneratedKeyWord(5));
@@ -100,21 +106,21 @@ public class PasswordResetServiceImpl implements PasswordResetService{
     @Transactional(readOnly = true)
     public boolean isValidUserKeyWord(PasswordResetData passwordResetData) {
         if(TextUtils.isEmpty(passwordResetData.getLogin()))
-            throw ExceptionFactory.create(PasswordResetError.EMPTY_LOGIN);
+            throw ExceptionFactory.create(InputError.EMPTY_LOGIN);
         if(TextUtils.isEmpty(passwordResetData.getKeyWord()))
-            throw ExceptionFactory.create(PasswordResetError.EMPTY_KEY_WORD);
+            throw ExceptionFactory.create(InputError.EMPTY_KEY_WORD);
 
         PasswordResetKeyWord keyWordFromDB = keyWordRepository.findByKeyWord(passwordResetData.getKeyWord().toUpperCase());
-        if(keyWordFromDB.isExpired()){
+        if(keyWordFromDB != null && keyWordFromDB.isExpired()){
             keyWordRepository.delete(keyWordFromDB);
             throw ExceptionFactory.create(PasswordResetError.KEY_WORD_IS_EXPIRED);
         }
 
-        User keyWordUser = keyWordRepository.findUserByKeyWord(passwordResetData.getKeyWord().toUpperCase());
         User user = getUserByLoginOrEmail(passwordResetData.getLogin());
         if(user == null)
-            throw ExceptionFactory.create(PasswordResetError.NO_SUCH_USER);
+            throw ExceptionFactory.create(NotFoundError.NO_SUCH_USER);
 
+        User keyWordUser = keyWordRepository.findUserByKeyWord(passwordResetData.getKeyWord().toUpperCase());
         if(keyWordUser == null)
             return false;
 
@@ -127,23 +133,19 @@ public class PasswordResetServiceImpl implements PasswordResetService{
     @Transactional
     public void resetUserPassword(PasswordResetData passwordResetData) {
         if(TextUtils.isEmpty(passwordResetData.getLogin()))
-            throw ExceptionFactory.create(PasswordResetError.EMPTY_LOGIN);
+            throw ExceptionFactory.create(InputError.EMPTY_LOGIN);
         if(TextUtils.isEmpty(passwordResetData.getKeyWord()))
-            throw ExceptionFactory.create(PasswordResetError.EMPTY_KEY_WORD);
+            throw ExceptionFactory.create(InputError.EMPTY_KEY_WORD);
         if(TextUtils.isEmpty(passwordResetData.getPassword()))
-            throw ExceptionFactory.create(PasswordResetError.EMPTY_PASSWORD);
-
-        if(!isValidUserKeyWord(passwordResetData))
-            throw ExceptionFactory.create(PasswordResetError.INCORRECT_KEY_WORD);
+            throw ExceptionFactory.create(InputError.EMPTY_PASSWORD);
 
         Pattern pattern = Pattern.compile(PASSWORD_REGEX_PATTERN);
         Matcher matcher = pattern.matcher(passwordResetData.getPassword());
         if (!matcher.find())
-            throw ExceptionFactory.create(PasswordResetError.INCORRECT_PASSWORD);
+            throw ExceptionFactory.create(InputError.INVALID_PASSWORD);
 
         User keyWordUser = keyWordRepository.findUserByKeyWord(passwordResetData.getKeyWord().toUpperCase());
-        keyWordUser.setPassword(passwordResetData.getPassword());
-        userRepository.updateUserPassword(keyWordUser.getUserId(), keyWordUser.getPassword());
+        userRepository.updateUserPassword(keyWordUser.getUserId(), passwordEncoder.encode(passwordResetData.getPassword()));
         keyWordRepository.deleteByUser_login(passwordResetData.getLogin());
     }
 
@@ -154,7 +156,7 @@ public class PasswordResetServiceImpl implements PasswordResetService{
             pattern = Pattern.compile(EMAIL_REGEX_PATTERN);
             matcher = pattern.matcher(login);
             if (!matcher.find())
-                throw ExceptionFactory.create(PasswordResetError.INCORRECT_LOGIN);
+                throw ExceptionFactory.create(InputError.INVALID_LOGIN);
             else
                 return userRepository.findUserByEmail(login);
         }
